+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.5.2 - 2026-03-08
+
+### Added
+
+- Added support for sending the battery level as a percentage via the `battery-percentage` header.
+
+### Fixed
+
+- Fixed the application of model bit depth when rendering.
+
 ## 0.5.1 - 2026-03-02
 
 ### Fixed